VeloxScan: Fix the computation of the pre-allocation size of leaf buffers (#233)
Previously, the `rapids-hybrid-execution` module reworked the
computation logic about pre-allocation size of leaf buffers: no longer
estimated the pre-allocation size by the last batch, instead estimated
based on the accumulated size of all processed batches.

However, this commit also introduced a bug: when there are enormous
batches being processed, the accumulated `dataSize` or `offsetSize` may
overflow the INT_32, should have used INT_64.

This PR is to fix this bug which will be triggered under extreme large
data scenario. And this PR does not refer any PR of the public repo,
since this PR is copied from the `rapids-hybrid-execution` repo.

+++ rapids-velox/src/main/scala/com/nvidia/spark/rapids/velox/SerializableMessages.scala
@@ -106,13 +106,14 @@ object ConverterMetrics extends SerializableMessage {
  * Data class holds the decoded sample information of a specific field.
  * posInSchema means the position of this field in the preorder flattened schema.
  */
-case class SampleColumnInfo(posInSchema: Int,
-                            veloxType: VeloxDataType,
-                            readType: StructField,
-                            numRows: Int,
-                            offsetsSize: Int,
-                            dataSize: Int,
-                            children: Seq[SampleColumnInfo])
+case class SampleColumnInfo(
+    posInSchema: Int,
+    veloxType: VeloxDataType,
+    readType: StructField,
+    numRows: Long,
+    offsetsSize: Long,
+    dataSize: Long,
+    children: Seq[SampleColumnInfo])
 
 /**
  * The message serialized by the backend method `encodeSampleInfo`, which carries buffer sizes
@@ -158,11 +159,10 @@ object SampleColumnMsg extends SerializableMessage {
         posInSchema = helper.head,
         veloxType = VeloxDataType.decodeVeloxType(msg(offset + 1).toInt),
         readType = helper.targetType,
-        numRows = msg(offset + 2).toInt,
-        offsetsSize = msg(offset + 3).toInt,
-        dataSize = msg(offset + 4).toInt,
-        children = children
-      )
+        numRows = msg(offset + 2),
+        offsetsSize = msg(offset + 3),
+        dataSize = msg(offset + 4),
+        children = children)
     }
 
     val stack = mutable.Stack[DecodeHelper]()
